go.crypto/ssh/test: don't kill process if it was never started
This case arises if s.cmd.Start fails when called by
server.Dial.

R=golang-dev, dave
CC=golang-dev
https://golang.org/cl/6821097

+++ ssh/test/test_unix_test.go
@@ -193,8 +193,10 @@ func (s *server) Dial(config *ssh.ClientConfig) *ssh.ClientConn {
 }
 
 func (s *server) Shutdown() {
-	if err := s.cmd.Process.Kill(); err != nil {
-		s.t.Error(err)
+	if s.cmd.Process != nil {
+		if err := s.cmd.Process.Kill(); err != nil {
+			s.t.Error(err)
+		}
 	}
 	if s.t.Failed() {
 		// log any output from sshd process
